requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
annotated-types==0.7.0
assertpy==1.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
curlify==3.0.0
distconfig3==1.0.1
exceptiongroup==1.3.0
Faker==37.8.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyHamcrest==2.1.0
pytest==8.4.1
PyYAML==6.0.2
requests==2.32.4
retrying==1.4.2
structlog==25.4.0
toml==0.10.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vyper-config==1.2.1
watchdog==6.0.0
swagger-coverage @https://github.com/tatvitus/swagger-coverage-py/archive/refs/heads/master.zip
